[ADD] Make Square; move extensions to their own class
