+++ Assets/Lightspeed/Unity/Helpers/Utility.Textures.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Experimental.Rendering;
@@ -6,8 +7,6 @@ namespace Rhinox.Lightspeed
 {
     public static partial class Utility
     {
-        private static Dictionary<Color, Texture2D> _textureByColor;
-
         /// <summary>
         /// Copies a texture (readable or not) into CPU memory of a new texture.
         /// Apply needs to be called for it to be usable in GPU.
@@ -29,6 +28,8 @@ namespace Rhinox.Lightspeed
             return copy;
         }
         
+        private static Dictionary<Color, Texture2D> _textureByColor;
+
         public static Texture2D GetColorTexture(Color color)
         {
             if (_textureByColor == null)
@@ -43,73 +44,6 @@ namespace Rhinox.Lightspeed
             }
 
             return _textureByColor[color];
-        }
-
-        public static Texture2D Pad(this Texture2D tex, int padding)
-            => Pad(tex, new RectOffset(padding, padding, padding, padding));
-        
-        public static Texture2D Pad(this Texture2D tex, Vector2Int padding)
-            => Pad(tex, new RectOffset(padding.x, padding.y, padding.x, padding.y));
-
-        public static Texture2D Pad(this Texture2D tex, RectOffset padding)
-        {
-            if (tex == null)
-                return tex;
-            
-            Texture2D paddedTex = new Texture2D(1, 1, tex.format, tex.mipmapCount > 0) {
-#if UNITY_EDITOR
-                alphaIsTransparency = tex.alphaIsTransparency,
-#endif
-                wrapMode = tex.wrapMode,
-                wrapModeU = tex.wrapModeU,
-                wrapModeV = tex.wrapModeV,
-                wrapModeW = tex.wrapModeW,
-                filterMode = tex.filterMode,
-                anisoLevel = tex.anisoLevel
-            };
-
-            var color = Color.clear;
-            paddedTex.SetPixel(0, 0, color);
-            paddedTex.Resize(tex.width + padding.left + padding.right, tex.height + padding.top + padding.bottom);
-            paddedTex.Apply();
-            
-            Graphics.CopyTexture(
-                tex, 0, 0, 0, 0, tex.width, tex.height,
-                paddedTex, 0, 0, padding.left, padding.top
-            );
-
-            // Not sure WHY this is needed but unfortunately the color does not properly scale (?)
-            InsetBorder(paddedTex, padding, color);
-
-            paddedTex.Apply(true);
-            return paddedTex;
-        }
-
-        public static void InsetBorder(this Texture2D tex, RectOffset border, Color color)
-        {
-            for (int x = 0; x < border.left; ++x)
-            {
-                for (int y = 0; y < tex.height; ++y)
-                    tex.SetPixel(x, y, color);
-            }
-
-            for (int x = 0; x < border.right; ++x)
-            {
-                for (int y = 0; y < tex.height; ++y)
-                    tex.SetPixel(tex.width - 1 - x, y, color);
-            }
-
-            for (int y = 0; y < border.bottom; ++y)
-            {
-                for (int x = border.left; x < tex.width - border.right; ++x)
-                    tex.SetPixel(x, y, color);
-            }
-
-            for (int y = 0; y < border.top; ++y)
-            {
-                for (int x = border.left; x < tex.width - border.right; ++x)
-                    tex.SetPixel(x, tex.height - 1 - y, color);
-            }
         }
     }
 }
